+++ era-solc/src/standard_json/input/settings/error_type.rs
@@ -12,6 +12,8 @@ use std::str::FromStr;
 pub enum ErrorType {
     /// The eponymous feature.
     SendTransfer,
+    /// The eponymous feature.
+    Ripemd160,
 }
 
 impl ErrorType {
@@ -32,6 +34,7 @@ impl FromStr for ErrorType {
     fn from_str(string: &str) -> Result<Self, Self::Err> {
         match string {
             "sendtransfer" => Ok(Self::SendTransfer),
+            "ripemd160" => Ok(Self::Ripemd160),
             r#type => Err(anyhow::anyhow!("Invalid suppressed error type: {type}")),
         }
     }
@@ -41,6 +44,7 @@ impl std::fmt::Display for ErrorType {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         match self {
             Self::SendTransfer => write!(f, "sendtransfer"),
+            Self::Ripemd160 => write!(f, "ripemd160"),
         }
     }
 }