+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # The `zksolc` changelog
 
+## [Unreleased]
+
+### Added
+
+- An AST error for the usage of `ripemd160` precompile
+
+### Changed
+
+- The warning about the default codegen is made more verbose and informative
+- Updated to Rust v1.83.0
+
 ## [1.5.11] - 2025-01-27
 
 ### Fixed